# service/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SwarX API",
    "version": "1.0.0",
    "description": "API documentation for SwarX"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Code",
        "description": "Access token for API authentication"
      }
    }
  },
  "paths": {
...
    "/songs": {
      "get": {
        "summary": "Get songs by IDs or link",
        "description": "Retrieve songs using either a comma-separated list of song IDs or a JioSaavn song link",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of song IDs"
          },
          {
            "in": "query",
            "name": "link",
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "JioSaavn song URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved songs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Song"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          }
        }
      }
    },
    "/songs/{id}": {
      "get": {
        "summary": "Get song by ID",
        "description": "Retrieve detailed information about a specific song, optionally including lyrics",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          },
          {
            "in": "query",
            "name": "lyrics",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Include lyrics in the response"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved song details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Song"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "404": {
            "description": "Song not found"
          }
        }
      }
    },
    "/songs/{id}/lyrics": {
      "get": {
        "summary": "Get song lyrics",
        "description": "Retrieve lyrics for a specific song",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved song lyrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "lyrics": {
                          "type": "string"
                        },
                        "snippetLyrics": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "404": {
            "description": "Lyrics not found"
          }
        }
      }
    },
    "/songs/{id}/suggestions": {
      "get": {
        "summary": "Get song suggestions",
        "description": "Get suggested songs based on a specific song",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "description": "Maximum number of suggestions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved song suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Song"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "404": {
            "description": "Song not found"
          }
        }
      }
    }
  },
...
}